requirements.txt
python-dotenv==1.0.0
weaviate-client>=3.26.7,<4.0.0
requests>=2.32.2,<3.0.0
langchain>=0.2.6,<0.3.0
langchain-ollama>=0.1.3,<0.2.0
pypdf==3.17.0
flask>=2.3.0,<3.0.0
hvac>=1.0.0
sqlalchemy>=2.0.0,<3.0.0
psycopg2-binary>=2.9.0,<3.0.0